# openapi.yaml
openapi: 3.1.0
info:
  title: Schedy
  version: 1.0.0
  summary: Self-hostable HTTP task scheduler.
  description: >-
    Schedy is a self-hostable HTTP task scheduler written in Go. Schedule an
    HTTP request to be delivered to any URL at a future time, with optional
    retries, fixed or exponential backoff, and interval-based recurrence.


    Authentication is via an API key sent in the `X-API-Key` header. The key is
    enforced only when the server is started with the `SCHEDY_API_KEY`
    environment variable set; when it is unset the server accepts anonymous
    requests, but the security scheme still logically applies to the `/tasks`
    and `/admin` routes. A request missing a required key returns `401`; a
    request carrying an invalid key returns `403`.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
...
servers:
  - url: http://localhost:8080
    description: Local Schedy instance
...
paths:
  /tasks:
    post:
      tags:
        - Tasks
      operationId: createTask
      summary: Schedule a new task
      description: >-
        Schedule a new HTTP task for a future time. The `execute_at` timestamp
        must be in the future. Supplying an optional `Idempotency-Key` request
        header makes the create idempotent: a repeat of a request whose key
        already matches a pending task returns that existing task with a `200`
        instead of creating a new one. Without a key, an identical schedule
        (same `url` and `execute_at` within one second) is treated as a
        duplicate.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: >-
            Caller-supplied key that makes creation idempotent. A repeat request
            with the same key returns the previously created pending task.
          schema:
            type: string
          example: reminder-42-20300101
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/taskRequest'
            example:
              url: https://api.example.com/webhooks/reminder
              method: POST
              headers:
                Content-Type: application/json
                X-Source: schedy
              payload:
                message: Time to stand up
                user_id: 42
              execute_at: "2030-01-01T09:00:00Z"
              retries: 3
              retry_interval: 5000
              retry_mode: exponential
              schedule: "24h"
      responses:
        '201':
          description: Task created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              example:
                id: d290f1ee-6c54-4b01-90e6-d701748f0851
                idempotency_key: reminder-42-20300101
                url: https://api.example.com/webhooks/reminder
                method: POST
                execute_at: "2030-01-01T09:00:00Z"
                headers:
                  Content-Type: application/json
                  X-Source: schedy
                payload:
                  message: Time to stand up
                  user_id: 42
                retries: 3
                retry_interval: 5000
                retry_mode: exponential
                schedule: "24h"
                status: pending
        '200':
          description: >-
            An idempotent or duplicate match was found; the existing pending
            task is returned unchanged.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              example:
                id: d290f1ee-6c54-4b01-90e6-d701748f0851
                idempotency_key: reminder-42-20300101
                url: https://api.example.com/webhooks/reminder
                method: POST
                execute_at: "2030-01-01T09:00:00Z"
                headers:
                  Content-Type: application/json
                  X-Source: schedy
                payload:
                  message: Time to stand up
                  user_id: 42
                retries: 3
                retry_interval: 5000
                retry_mode: exponential
                schedule: "24h"
                status: pending
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
    get:
      tags:
        - Tasks
      operationId: listTasks
      summary: List tasks
      description: >-
        List one page of scheduled tasks, optionally filtered by lifecycle
        status. Results are paginated: read `next_cursor` from the response and
        pass it back as `cursor` to fetch the following page, repeating while
        `has_more` is true. The cursor is opaque - do not construct or parse it -
        and it is only valid for the same `status` filter that produced it.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: status
          in: query
          required: false
          description: Filter tasks by lifecycle status.
          schema:
            type: string
            enum:
              - pending
              - running
              - succeeded
              - failed
              - cancelled
        - name: limit
          in: query
          required: false
          description: Maximum number of tasks to return in this page.
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: cursor
          in: query
          required: false
          description: >-
            Opaque cursor from a previous response's `next_cursor`. Omit for the
            first page.
          schema:
            type: string
      responses:
        '200':
          description: One page of tasks matching the filter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskPage'
              example:
                tasks:
                  - id: d290f1ee-6c54-4b01-90e6-d701748f0851
                    url: https://api.example.com/webhooks/reminder
                    method: POST
                    execute_at: "2030-01-01T09:00:00Z"
                    headers:
                      X-Source: schedy
                    payload:
                      message: Time to stand up
                    retries: 3
                    retry_interval: 5000
                    retry_mode: exponential
                    status: pending
                next_cursor: dGFzazpwZW5kaW5nOjAwMDAwMDAxODkzNDgzMjAwOmQyOTBmMWVl
                has_more: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags:
        - Tasks
      operationId: deleteTasks
      summary: Bulk-delete tasks
      description: >-
        Delete tasks matching one or more filters. At least one of `url`,
        `before`, or `after` must be supplied, otherwise the request is
        rejected with `400`.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: url
          in: query
          required: false
          description: Delete tasks whose target URL matches this value.
          schema:
            type: string
        - name: before
          in: query
          required: false
          description: Delete tasks scheduled before this RFC3339 timestamp.
          schema:
            type: string
            format: date-time
        - name: after
          in: query
          required: false
          description: Delete tasks scheduled after this RFC3339 timestamp.
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: The number of tasks deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkDeleteResponse'
              example:
                deleted: 3
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Unique identifier of the task.
        schema:
          type: string
        example: d290f1ee-6c54-4b01-90e6-d701748f0851
    get:
      tags:
        - Tasks
      operationId: getTask
      summary: Get a task by ID
      description: Retrieve a single task, including its attempt history.
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: The requested task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags:
        - Tasks
      operationId: updateTask
      summary: Update a pending task
      description: >-
        Replace the client-owned fields of a task while keeping its ID. Only
        tasks in the `pending` state are mutable; updating a task in any other
        state returns `409`. Server-owned state (status, attempts, finished_at)
        is preserved.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/taskRequest'
            example:
              url: https://api.example.com/webhooks/reminder
              method: POST
              headers:
                X-Source: schedy
              payload:
                message: Time to stand up (rescheduled)
              execute_at: "2030-01-01T10:00:00Z"
              retries: 5
              retry_interval: 3000
              retry_mode: fixed
      responses:
        '200':
          description: The updated task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags:
        - Tasks
      operationId: cancelTask
      summary: Cancel a task
      description: >-
        Cancel a task by ID. Non-terminal tasks are soft-cancelled (marked
        `cancelled` and retained in history); already-terminal tasks are a
        no-op. Returns no content on success.
      security:
        - ApiKeyAuth: []
      responses:
        '204':
          description: The task was cancelled (or was already terminal). No content.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /tasks/{id}/run:
    parameters:
      - name: id
        in: path
        required: true
        description: Unique identifier of the task.
        schema:
          type: string
        example: d290f1ee-6c54-4b01-90e6-d701748f0851
    post:
      tags:
        - Tasks
      operationId: replayTask
      summary: Replay a task
      description: >-
        Re-arm a finished task: same id, back to `pending`, `execute_at` set to
        now, picked up by the runner within a few seconds. The attempt log is
        kept rather than cleared - the delivery that failed is the reason for
        replaying - so a replay appends to it. `finished_at` is cleared and set
        again when the replay finishes, and the task's retry budget applies
        afresh. Only terminal tasks (failed, succeeded, cancelled) may be
        replayed; a pending task is already scheduled and a running one is
        mid-delivery, so re-arming either would race the runner. Note that
        replaying a succeeded task delivers its payload a second time.
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: The re-armed task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              example:
                id: d290f1ee-6c54-4b01-90e6-d701748f0851
                url: https://api.example.com/webhooks/reminder
                method: POST
                execute_at: "2030-01-01T15:04:05Z"
                headers:
                  X-Source: schedy
                payload:
                  message: Time to stand up
                retries: 3
                retry_interval: 5000
                retry_mode: exponential
                status: pending
                attempts:
                  - n: 1
                    fired_at: "2030-01-01T09:00:00Z"
                    status_code: 500
                    error: unexpected status 500
                    duration_ms: 42
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: >-
            The task is pending or running, so it is not eligible for replay.
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        API key sent in the `X-API-Key` header. Enforced only when the server
        is started with the `SCHEDY_API_KEY` environment variable set; when
        unset the server accepts anonymous requests but the scheme still applies
        to these routes. A missing key returns `401`; an invalid key returns
        `403`.
  responses:
    BadRequest:
      description: >-
        The request failed validation - for example a missing `url`, an invalid
        `method`, a malformed or past `execute_at`, an invalid `retry_mode`, an
        invalid `schedule`, an unparseable time filter, or a bulk delete with no
        filter. The body is a plain-text reason.
      content:
        text/plain:
          schema:
            type: string
          example: "url is required\n"
...
  schemas:
...
    taskRequest:
      type: object
      description: >-
        The client-owned shape of a task, shared by create and update.
        Server-owned state (id, status, attempts, finished_at) is not accepted.
      required:
        - url
        - execute_at
      properties:
        url:
          type: string
          description: Absolute URL the scheduled HTTP request is delivered to.
          example: https://api.example.com/webhooks/reminder
        execute_at:
          type: string
          format: date-time
          description: >-
            RFC3339 timestamp for when the task should fire. Must be in the
            future at the time of the request.
          example: "2030-01-01T09:00:00Z"
        method:
          type: string
          enum:
            - GET
            - POST
            - PUT
            - PATCH
            - DELETE
            - HEAD
          default: POST
          description: HTTP verb used to deliver the task. Case-insensitive; defaults to POST.
        headers:
          type: object
          additionalProperties:
            type: string
          description: Custom HTTP headers sent with the delivered request.
        payload:
          description: >-
            Arbitrary request payload delivered as the body. May be any JSON
            value - object, array, string, number, boolean, or null.
        retries:
          type: integer
          default: 0
          description: Number of retries attempted after the first delivery fails.
        retry_interval:
          type: integer
          default: 2000
          description: Base delay between retries, in milliseconds.
        retry_mode:
          type: string
          enum:
            - fixed
            - exponential
          default: fixed
          description: >-
            How the delay between retries is computed. `fixed` waits
            `retry_interval` between every attempt; `exponential` backs off with
            full jitter.
        schedule:
          type: string
          description: >-
            Optional Go duration string (e.g. "15m", "2h") that makes the task
            recurring - after each fire a fresh task is enqueued at
            fire_time + schedule. Must be positive. Interval-based only; cron
            expressions are not supported.
          example: "24h"
    Task:
      type: object
      description: A scheduled task together with its current execution state.
      required:
        - id
        - url
        - method
        - execute_at
        - headers
        - payload
        - retries
        - retry_interval
        - retry_mode
        - status
      properties:
        id:
          type: string
          description: Server-generated unique identifier (UUID).
        idempotency_key:
          type: string
          description: >-
            The `Idempotency-Key` the task was created with, if any. Set once at
            creation and never changed.
        url:
          type: string
          description: Absolute URL the request is delivered to.
        method:
          type: string
          enum:
            - GET
            - POST
            - PUT
            - PATCH
            - DELETE
            - HEAD
          description: HTTP verb used to deliver the task.
        execute_at:
          type: string
          format: date-time
          description: RFC3339 timestamp for when the task fires.
        headers:
          type: object
          additionalProperties:
            type: string
          description: Custom HTTP headers sent with the delivered request.
        payload:
          description: >-
            Arbitrary payload delivered as the request body. May be any JSON
            value - object, array, string, number, boolean, or null.
        retries:
          type: integer
          description: Number of retries attempted after the first delivery fails.
        retry_interval:
          type: integer
          description: Base delay between retries, in milliseconds.
        retry_mode:
          type: string
          enum:
            - fixed
            - exponential
          description: How the delay between retries is computed.
        schedule:
          type: string
          description: >-
            Go duration for recurrence, present only when the task is recurring.
        status:
          type: string
          enum:
            - pending
            - running
            - succeeded
            - failed
            - cancelled
          description: The lifecycle state of the task. Exactly one at any time.
        attempts:
          type: array
          description: The delivery attempts fired so far, present once any have run.
          items:
            $ref: '#/components/schemas/Attempt'
        finished_at:
          type:
            - string
            - "null"
          format: date-time
          description: >-
            When the task reached a terminal state. Null or absent while the
            task is not yet terminal.
      example:
        id: d290f1ee-6c54-4b01-90e6-d701748f0851
        idempotency_key: reminder-42-20300101
        url: https://api.example.com/webhooks/reminder
        method: POST
        execute_at: "2030-01-01T09:00:00Z"
        headers:
          X-Source: schedy
        payload:
          message: Time to stand up
        retries: 3
        retry_interval: 5000
        retry_mode: exponential
        status: succeeded
        attempts:
          - n: 1
            fired_at: "2030-01-01T09:00:01Z"
            status_code: 200
            duration_ms: 142
        finished_at: "2030-01-01T09:00:01Z"
...